requirements.txt
flask
pillow
numpy
gunicorn
